refactor(util) : 패키지명 변경 -> error
