+++ src/main/java/domain/AttendanceStatus.java
@@ -4,8 +4,8 @@ import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-import static util.ErrorMessage.CLOSED_DAY;
-import static util.ErrorMessage.CLOSED_TIME;
+import static error.ErrorMessage.CLOSED_DAY;
+import static error.ErrorMessage.CLOSED_TIME;
 import static java.time.DayOfWeek.SATURDAY;
 import static java.time.DayOfWeek.SUNDAY;
