+++ src/main/java/view/MenuOption.java
@@ -1,6 +1,6 @@
 package view;
 
-import util.ErrorMessage;
+import error.ErrorMessage;
 
 import java.util.Arrays;
