+++ src/test/java/AllCrewTest.java
@@ -1,4 +1,4 @@
-import util.ErrorMessage;
+import error.ErrorMessage;
 import domain.AllCrew;
 import domain.Attendance;
 import domain.Crew;
